chore: move MapView Header into own file
