+++ src/views/Map/styles.ts
@@ -1,7 +1,7 @@
 import { styled } from '@mui/material'
 
 export const IFrame = styled('iframe')`
-  width: 100vw;
+  width: 100%;
   height: calc(100vh - 80px);
   border: none;
 `
